Fb/server debug just in time (#115)
* just in time enablement

* just in time enablement 2

* record nock for srv 1

* record nock for srv 2

* remove start debugger command not needed

* srv debug test

* srv debug test 2

* less nock recordings

+++ src/submodules/serverDiagnostic.js
@@ -2,68 +2,20 @@
 
 const { orderedStringify, writeTextSync } = require("../shared/static");
 const { assert } = require("../shared/error");
-const { request } = require("../shared/request");
 const { Logger } = require("../shared/logger");
 
+const RUNTIME = {
+  NODE: "node",
+  JAVA: "java",
+};
+
 const BUILDPACK_INFO = {
-  nodejs_buildpack: { runtime: "node", debugPort: 9229 },
-  java_buildpack: { runtime: "java", debugPort: 8000 },
+  nodejs_buildpack: { runtime: RUNTIME.NODE, debugPort: 9229 },
+  java_buildpack: { runtime: RUNTIME.JAVA, debugPort: 8000 },
 };
 const DEFAULT_ENV_FILENAME = "default-env.json";
 
 const logger = Logger.getInstance();
-
-const _serverDebug = async (context, { appName, appInstance } = {}) => {
-  const { cfBuildpack, cfAppGuid, cfRouteUrl, cfSsh } = appName
-    ? await context.getAppNameInfoCached(appName)
-    : await context.getSrvInfo();
-  const cfBuildpackInfoKey =
-    cfBuildpack &&
-    Object.keys(BUILDPACK_INFO).find(
-      (cfBuildpackName) =>
-        cfBuildpack.includes(cfBuildpackName) || cfBuildpack.includes(cfBuildpackName.replace(/_/g, "-"))
-    );
-  const { runtime, debugPort: inferredPort } = (cfBuildpackInfoKey && BUILDPACK_INFO[cfBuildpackInfoKey]) || {};
-  const localPort = inferredPort || 8000;
-
-  let responseData = {};
-  if (cfRouteUrl) {
-    try {
-      const token = await context.getCachedUaaToken();
-      const response = await request({
-        url: cfRouteUrl,
-        pathname: "/info",
-        auth: { token },
-        headers: {
-          "X-Cf-App-Instance": `${cfAppGuid}:${appInstance}`,
-        },
-        logged: false,
-        checkStatus: false,
-      });
-
-      responseData = response.ok && (await response.json());
-    } catch (err) {} // eslint-disable-line no-empty
-  }
-  const { debugPort } = responseData;
-  const remotePort = debugPort || inferredPort;
-  assert(remotePort, `could not determine remote debugPort from /info or infer from buildpack`);
-
-  logger.info();
-  if (!debugPort) {
-    logger.info(
-      `could not determine remote debugPort from /info, falling back to ${cfBuildpack} default ${remotePort}`
-    );
-  }
-  logger.info(`connect ${runtime ? runtime + " debugger" : "debugger"} on port ${localPort}`);
-  logger.info(`use request header "X-Cf-App-Instance: ${cfAppGuid}:${appInstance}" to target this app instance`);
-  logger.info();
-  await cfSsh({ localPort, remotePort, appInstance });
-};
-
-const serverDebug = async (context, [appName, appInstance = "0"]) => {
-  assert(/\d+/.test(appInstance), `argument "${appInstance}" is not a valid app instance`);
-  return _serverDebug(context, { appName, appInstance });
-};
 
 const serverEnvironment = async (context, [appName]) => {
   const { cfEnvServices, cfEnvApp, cfEnvVariables } = appName
@@ -87,15 +39,42 @@ const serverCertificates = async (context, [appName, appInstance = "0"]) => {
   logger.info("saved instance certificates");
 };
 
-const serverStartDebugger = async (context, [appName, appInstance = "0"]) => {
+const _serverDebug = async (context, { appName, appInstance } = {}) => {
+  const { cfBuildpack, cfAppGuid, cfSsh } = appName
+    ? await context.getAppNameInfoCached(appName)
+    : await context.getSrvInfo();
+  const cfBuildpackInfoKey =
+    cfBuildpack &&
+    Object.keys(BUILDPACK_INFO).find(
+      (cfBuildpackName) =>
+        cfBuildpack.includes(cfBuildpackName) || cfBuildpack.includes(cfBuildpackName.replace(/_/g, "-"))
+    );
+  const { runtime, debugPort: inferredPort } = (cfBuildpackInfoKey && BUILDPACK_INFO[cfBuildpackInfoKey]) || {};
+  assert(inferredPort, `could not infer remote debugPort from buildpack "${cfBuildpack}"`);
+  const localPort = inferredPort;
+  const remotePort = inferredPort;
+
+  logger.info();
+  if (runtime === RUNTIME.NODE) {
+    try {
+      await cfSsh({ command: "pkill --signal SIGUSR1 node", appInstance });
+    } catch (err) {
+      logger.warning("warning: could not enable debugging for node process: ", err.message);
+    }
+  }
+  logger.info(`connect ${runtime ? runtime + " debugger" : "debugger"} on port ${localPort}`);
+  logger.info(`use request header "X-Cf-App-Instance: ${cfAppGuid}:${appInstance}" to target this app instance`);
+  logger.info();
+  await cfSsh({ localPort, remotePort, appInstance });
+};
+
+const serverDebug = async (context, [appName, appInstance = "0"]) => {
   assert(/\d+/.test(appInstance), `argument "${appInstance}" is not a valid app instance`);
-  const { cfSsh } = appName ? await context.getAppNameInfoCached(appName) : await context.getSrvInfo();
-  await cfSsh({ command: "pkill --signal SIGUSR1 node", appInstance });
+  return _serverDebug(context, { appName, appInstance });
 };
 
 module.exports = {
-  serverDebug,
   serverEnvironment,
   serverCertificates,
-  serverStartDebugger,
+  serverDebug,
 };
